+++ flake/default.nix
@@ -1,0 +1,12 @@
+{
+  systems = [
+    "x86_64-linux"
+    "aarch64-linux"
+    "x86_64-darwin"
+    "aarch64-darwin"
+  ];
+
+  imports = [
+    ./modules/zebra.nix
+  ];
+}